can insert link vertices

+++ src/clibs/prtzl.h
@@ -1,5 +1,6 @@
 
 #include "map.h"
+#include "list.h"
 
 struct graph{
 	struct map* vertices;
@@ -18,13 +19,13 @@ struct node* insert_vertex(struct graph*, char*);
 struct node* query_vertex(struct graph*, char*);
 
 /* <_ "key" _> */
-int delete_vertex(struct graph*, char*);
+double delete_vertex(struct graph*, char*);
 
 void put_node_property(struct node*, char*, void*);
 
 void* get_node_property(struct node*, char*);
 
-int link(struct node*, struct node*, double weight);
+double link(struct node*, struct node*, double weight);
 
-int bi_link(struct node*, struct node*, double weight);
+double bi_link(struct node*, struct node*, double weight);
